Move DirtyModel used for test into activerecord/test/models/dirty_model.rb
This may now be re-used in other tests.
